pyproject.toml
[project]
name = "scikit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyterlab>=4.5.3",
    "langdetect>=1.0.9",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.2",
]
